+++ src/_metronic/layout/components/header/Topbar.js
@@ -26,12 +26,6 @@ export function Topbar() {
 
   return (
     <div className="topbar">
-      {layoutProps.viewSearchDisplay && <SearchDropdown />}
-
-      {layoutProps.viewNotificationsDisplay && <UserNotificationsDropdown />}
-
-
-
       {layoutProps.viewUserDisplay && <QuickUserToggler />}
     </div>
   );
