+++ src/app/misc/enum.js
@@ -1,0 +1,6 @@
+ export var nvAcces = {
+    coordonnateur:1,
+    tuteur:2,
+    entreprise:3,
+    stagiaire:4
+}
