+++ src/app/modules/Auth/__mocks__/userTableMock.js
@@ -1,167 +1,56 @@
-import { toAbsoluteUrl } from "../../../../_metronic/_helpers";
+import { toAbsoluteUrl } from '../../../../_metronic/_helpers';
 
 export default [
   {
     id: 1,
-    username: "admin",
-    password: "demo",
-    email: "admin@demo.com",
+    username: "hcloutier",
+    password: "Admin123!",
+    email: "hugo.cloutier@cegeptr.qc.ca",
     accessToken: "access-token-8f3ae836da744329a6f93bf20594b5cc",
     refreshToken: "access-token-f8c137a2c98743f48b643e71161d90aa",
-    roles: [1], // Administrator
+    roles: [1], // coordonnateur
     pic: toAbsoluteUrl("/media/users/300_21.jpg"),
-    fullname: "Sean S",
-    firstname: "Sean",
-    lastname: "Stark",
-    occupation: "CEO",
-    companyName: "Keenthemes",
-    phone: "456669067890",
-    language: "en",
-    timeZone: "International Date Line West",
-    website: "https://keenthemes.com",
-    emailSettings: {
-      emailNotification: true,
-      sendCopyToPersonalEmail: false,
-      activityRelatesEmail: {
-        youHaveNewNotifications: false,
-        youAreSentADirectMessage: false,
-        someoneAddsYouAsAsAConnection: true,
-        uponNewOrder: false,
-        newMembershipApproval: false,
-        memberRegistration: true,
-      },
-      updatesFromKeenthemes: {
-        newsAboutKeenthemesProductsAndFeatureUpdates: false,
-        tipsOnGettingMoreOutOfKeen: false,
-        thingsYouMissedSindeYouLastLoggedIntoKeen: true,
-        newsAboutMetronicOnPartnerProductsAndOtherServices: true,
-        tipsOnMetronicBusinessProducts: true,
-      },
-    },
-    communication: {
-      email: true,
-      sms: true,
-      phone: false,
-    },
-    address: {
-      addressLine: "L-12-20 Vertex, Cybersquare",
-      city: "San Francisco",
-      state: "California",
-      postCode: "45000",
-    },
-    socialNetworks: {
-      linkedIn: "https://linkedin.com/admin",
-      facebook: "https://facebook.com/admin",
-      twitter: "https://twitter.com/admin",
-      instagram: "https://instagram.com/admin",
-    },
+    fullname: "Hugo Cloutier",
+    firstname: "Hugo",
+    lastname: "Cloutier",
   },
   {
     id: 2,
-    username: "user",
-    password: "demo",
-    email: "user@demo.com",
-    accessToken: "access-token-6829bba69dd3421d8762-991e9e806dbf",
-    refreshToken: "access-token-f8e4c61a318e4d618b6c199ef96b9e55",
-    roles: [2], // Manager
+    username: "vbaylly",
+    password: "vincent",
+    email: "vincent@demo.com",
+    accessToken: "access-token-6829xxa69dd3421d8762-991e9e806dbf",
+    refreshToken: "access-token-f8e4c61a318g4d618b6c199ef96b9t55",
+    roles: [2], // tuteur
     pic: toAbsoluteUrl("/media/users/100_2.jpg"),
-    fullname: "Megan F",
-    firstname: "Megan",
-    lastname: "Fox",
-    occupation: "Deputy Head of Keenthemes in New York office",
-    companyName: "Keenthemes",
-    phone: "456669067891",
-    language: "en",
-    timeZone: "International Date Line West",
-    communication: {
-      email: true,
-      sms: true,
-      phone: false,
-    },
-    emailSettings: {
-      emailNotification: true,
-      sendCopyToPersonalEmail: false,
-      activityRelatesEmail: {
-        youHaveNewNotifications: false,
-        youAreSentADirectMessage: false,
-        someoneAddsYouAsAsAConnection: true,
-        uponNewOrder: false,
-        newMembershipApproval: false,
-        memberRegistration: true,
-      },
-      updatesFromKeenthemes: {
-        newsAboutKeenthemesProductsAndFeatureUpdates: false,
-        tipsOnGettingMoreOutOfKeen: false,
-        thingsYouMissedSindeYouLastLoggedIntoKeen: true,
-        newsAboutMetronicOnPartnerProductsAndOtherServices: true,
-        tipsOnMetronicBusinessProducts: true,
-      },
-    },
-    address: {
-      addressLine: "3487  Ingram Road",
-      city: "Greensboro",
-      state: "North Carolina",
-      postCode: "27409",
-    },
-    socialNetworks: {
-      linkedIn: "https://linkedin.com/user",
-      facebook: "https://facebook.com/user",
-      twitter: "https://twitter.com/user",
-      instagram: "https://instagram.com/user",
-    },
+    fullname: "Vincent Baylly",
+    firstname: "Vincent",
+    lastname: "Baylly",
   },
   {
     id: 3,
-    username: "guest",
-    password: "demo",
-    email: "guest@demo.com",
-    accessToken: "access-token-d2dff7b82f784de584b60964abbe45b9",
-    refreshToken: "access-token-c999ccfe74aa40d0aa1a64c5e620c1a5",
-    roles: [3], // Guest
+    username: "entrepriseXYZ",
+    password: "entreprise",
+    email: "entreprise@demo.com",
+    accessToken: "access-token-d2dff7b82f784rf584b60964abbe45b9",
+    refreshToken: "access-token-c999ccfe74aa40d0bb1a64c5e620c1a5",
+    roles: [3], // entreprise
     pic: toAbsoluteUrl("/media/users/default.jpg"),
     fullname: "Manu G",
     firstname: "Manu",
     lastname: "Ginobili",
-    occupation: "CFO",
-    companyName: "Keenthemes",
-    phone: "456669067892",
-    language: "en",
-    timeZone: "International Date Line West",
-    communication: {
-      email: true,
-      sms: true,
-      phone: false,
-    },
-    emailSettings: {
-      emailNotification: true,
-      sendCopyToPersonalEmail: false,
-      activityRelatesEmail: {
-        youHaveNewNotifications: false,
-        youAreSentADirectMessage: false,
-        someoneAddsYouAsAsAConnection: true,
-        uponNewOrder: false,
-        newMembershipApproval: false,
-        memberRegistration: true,
-      },
-      updatesFromKeenthemes: {
-        newsAboutKeenthemesProductsAndFeatureUpdates: false,
-        tipsOnGettingMoreOutOfKeen: false,
-        thingsYouMissedSindeYouLastLoggedIntoKeen: true,
-        newsAboutMetronicOnPartnerProductsAndOtherServices: true,
-        tipsOnMetronicBusinessProducts: true,
-      },
-    },
-    address: {
-      addressLine: "1467  Griffin Street",
-      city: "Phoenix",
-      state: "Arizona",
-      postCode: "85012",
-    },
-    socialNetworks: {
-      linkedIn: "https://linkedin.com/guest",
-      facebook: "https://facebook.com/guest",
-      twitter: "https://twitter.com/guest",
-      instagram: "https://instagram.com/guest",
-    },
+  },
+  {
+    id: 4,
+    username: "emartins",
+    password: "stagiaire",
+    email: "ericmartins81@gmail.com",
+    accessToken: "access-token-8f3ae836da744329a6f93bf20594b5bb",
+    refreshToken: "access-token-f8c137a2c98743f48b643e71161d90cc",
+    roles: [4], // stagiaire
+    pic: toAbsoluteUrl("/media/users/default.jpg"),
+    fullname: "Eric Martins",
+    firstname: "Eric",
+    lastname: "Martins",
   },
 ];
